app/python/requirement.txt
click==8.1.7
dlib==19.24.4
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.26.4
opencv-python==4.10.0.84
pillow==10.4.0
